rename `views.py` to `views/api.py`
